requirements.txt
streamlit>=1.38.0
anthropic>=0.39.0
python-docx>=1.1.0
pydantic>=2.0.0
jinja2>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pillow>=10.0.0
numpy>=1.24.0
huggingface_hub>=0.20.0
# libsql-experimental>=0.0.40  # Only needed for Turso production mode; requires Rust to build
